+++ conkit/io/tests/test__cache.py
@@ -5,41 +5,40 @@ __date__ = "19 Jun 2017"
 
 import unittest
 
-from conkit.io._cache import _ParserCache
+from conkit.io._cache import ParserCache
 
 
-class Test_ParserCache(unittest.TestCase):
-
+class TestParserCache(unittest.TestCase):
     def test_1(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertTrue("casprr" in c)
 
     def test_2(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertFalse("foo" in c)
 
     def test_3(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertTrue("casprr" in c.contact_file_parsers)
 
     def test_4(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertFalse("bar" in c.contact_file_parsers)
 
     def test_5(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertTrue("fasta" in c.sequence_file_parsers)
 
     def test_6(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertFalse("baz" in c.sequence_file_parsers)
 
     def test_7(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertFalse("fasta" in c.contact_file_parsers)
 
     def test_8(self):
-        c = _ParserCache()
+        c = ParserCache()
         self.assertFalse("casprr" in c.sequence_file_parsers)